requirements.txt
torch
torchvision
torchaudio
diffusers
lightning
transformers
datasets
numpy
pandas
scipy
matplotlib
seaborn
notebook
torchmetrics
torcheval
accelerate
torch_fidelity
Flask
Pillow